Fix hotkeysGobal to work in input/textarea/select/checkbox
Setting hotkeysGlobal didn't allow ctrl+space or any other hotkey to
work while I was focused on an input.

The documentation for scopes in hotkey.js wasn't clear. After trying
to work through the code with the debugger I was none the wiser. How
are scopes supposed to be used and why do they exist?

I simply changed the call to: hotkeys.setScope("all"). I don't know if
this will cause issues, but I changed cp-advanced/index.html to have
the 4 input types for testing and I didn't see any issues. ctrl+space
plus ctrl+1 worked to activate the command palette and toggle the
theme.

Given the static value for setScope, you can probably remove the
tagName setting as well. However, this is all dark code for me so ....