add an option noButton to disable the mobile button
The button isn't really that useful if you are on a desktop system.
The shortcut key is easily accessible and the button is just clutter.
So add an option to suppress the mobile button when starting up command-pal.

This could also be useful if you have multiple command palettes. The
main one has a command button and other ones are invoked from
the main one or via their own hotkey. (I am not sure that this would
work but...).

Side note, modifying the CSS via JavaScript to include:

   button.mobile-button { display: none; }

possibly with a `!important` will also hide the button. This open
prevents the HTML for the button from being generated.