add HTML id's to selected elements to aid CSS styling
Different pages may have different styling requirements for command-pal
instances.

This commit adds id's to the main div (CommandPalette), the mobile
button (CommandPalette-button) and the modal-mask div
(CommandPalette-mask):. It also adds the option paletteId to replace
CommandPalette in the ids to provide unique ids and prevent
collisions.

It also alphabetizes the props passed to App in main.js.