add cp-placeholder class for svg and add doc
add the class cp-placeholder to the svg placeholder inserted when the
icon property is not defined. This allows commands without icon
properties to be flush left in the listbox. Just like using icon: null
on every command, but without having to do all the extra typing.

Added documentation for using icon.

Also updated code to use the "if ( !!var)" idiom.