bug: using enter to select menu item now clears search box
Type some characters to select a command and then type enter to
execute it. The palette closes. When you re-open the palette, the
search term you typed is still present and filtering the command list.

I expect the search term to be cleared similar to aborting with escape.

I am not sure if this is intentional or not, but I consider it a bug.
I could be convinced it is a feature to allow power users to repeat the
command by hitting return, open palette, hit return again etc. If so it needs to be documented as such.