add aliases search array and display best matches for aliases
Note: This upgrades the required version of fuse.js to 6.6.2. This
version allows weighting to be properly specified.

This adds an aliases array for each command. A match in the alias has
a weight of 2x compared to the name or description fields. If the user
enables the displayHints option, the best matching alias will be
appended to the end of the command name. This is useful for listing
alternate terms for the one that is chosen as the command
name. For example:

  name: emacs, alias: editor

will show up as emacs (editor) if the user typed edit.

The README.md has more info.

This was inspired by the implementation of synonyms described at:

https://blog.superhuman.com/how-to-build-a-remarkable-command-palette/#rule-4-build-flexibility-into-your-command-palette