handle long command names better; don't break layout.
Changed the height: 36px of the command list item paragraph to be a
min-height: 36px. If the command text gets too long, it is bleeding
over into the other command list rows.

I also set overflow-y: hidden in case somebody adds a max-height or
height to it again. This will stop the bleed over.

Also set flex-shrink:0 on the things that we expect in the icon area.
Also explicity target '.item :first-child' to prevent flex shrinkage
(a bit of defensive css):. Otherwise large command strings squish the
icons in column 1.