Make hotkey toggle CommandPal on/off
Make the hotkey able to toggle command pal on and dismiss it as well.

Inspired by

https://plutoapp.xyz/blog/post/designing-a-command-palette/#togglable

from: https://commandpalette.org

Basic reason is if a user accidently hits it, they can hit it again to
get out of it. Otherwise they have to figure out how to dismiss it.

Three changes:

1) add data-id=cp-SearchField to search input. When I first coded this
   it only worked with hotkeysGlobal: true. It would not work with
   hotkeysglobal: false.

   It finally dawned on me that the focus was in the search input when
   I sent the dismissal hotkey. hotkeys.js ignores hotkeys in an input
   by default. To fix this I had to find a way to uniquely identify
   this input to I could make hotkeys active. The normal id is a
   unique random string, so no joy there.

2) changed onMount to look at current state of showModal and call
   onClosed if true.

3) rewrote the hotkeys.js filter. Now I understand it. Wish I didn't
   8-).  When the filter function is defined, I make it a closure over
   BindToInputsToo. The new filter does 1 of three things:

      * if BindToInputsToo is true, it return true, so hotkeys will process it
      * if BindToInputsToo is false, check to see what the activeElement is.
        If its data-id is cp-SearchField return true so keys is processed.
      * otherwise look to see if it's a ContentEditable area, or
        a writable input or select. If so return false, else return true

Also I used https://github.com/jaywcjlove/hotkeys/issues/321 to
figure out how to create a filter, it's not correct.

  tagname = target.tagName

not

  tagname = target

I don't know if that's why hotkeysGlobal didn't work originally.

Also the default scope is all. So no need to set it.

All that matters is that the filter returns true to process the key or
false to ignore it. Setting the scope in the filter probably changes
the lookup table used to find matching keys.