fix ESC, arrow keys not working on mobile
This at east fixes keyboard keys ESC and arrow up/down on android
with chrome using the Hacker's keyboard.

From:

   https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/code

e.code() has issues with mapping the proper keyboard layout. For this
application with a physical keyboard and the keys being used I don't
think it matters but, if we wanted to intercept Y on an QWERTY keyboard
it would intercept Z on a QUERTZ layout.