able to exit from palette using 1 or more backspaces on empty search
On mobile, there is no good way to exit from the pallete using the
keyboard. Most mobile keyboards don't have ESC keys, or they are on a
secondary keyboard layout. The backspace key however is usually availble
on the primary keyboard layout.

This patch adds the ability to configure the number of consecutive
backspaces to exit the palette. The backspaces only count if the
search input is empty. If the number of required backspaces is > 1,
the number of remaining backspaces is shown if it is less than 5.
This feedback is generated by a template string. i18n may be an
issue. YAGNI says not to fix this now 8-) (if I even has a clue as to
how.)

Docs updated no examples changed. Works in Firefox and Chrome mobile
and desktop.

NOTE: this also patches the onKeyDown in SearchField to use e.key
rather than e.code otherwise it doesn't work on mobile where it is
most useful.