Added orderedCommands to select fuse.js or replacement sort function.
I realized this patch replaced the standard fuse.js sort bunction. It
is now an option to replace the sort function. Setting
orderedCommands: true replced the default fuse.js sort function.

This also includes a new replacement sort which is better than the
original.

The original would sort two items with scores of 0.03 and 2.2E-10 by
their order in the commands array. The new function sorts by score if
the score is < 0.009. For results that are close (< 0.05 apart), the
results will be sorted by their position in the command array. This
allows the dev to nudge results when the confidence in the scores is
lower.

This also pushes the debugOutput option down into App.svelte.

In README.md, I moved the aliases and hints section lower and added a
section on ordering matches.