Restore focus when clicking on button too. Report missing focus.
Activating the palette using the mobile button doesn't set focusedElement
from document.activeElement.

Even if it did, the button becomes the activeElement. Not exactly useful.

Set up a focus handler that looks at event.relatedTarget and use the
relatedTarget as the focusedElement if it can be focused. If
relatedTarget is null, use document.body.

also if focusedElement is falsy, console.error it.

Also added a period at the end of the title for the button in MobileButton.
It makes voice reading work better.